+++ include/measurement_kit/ndt/error.hpp
@@ -51,6 +51,9 @@ MK_DEFINE_ERR(MK_ERR_NDT(38), SerializingTestMsgError, "")
 MK_DEFINE_ERR(MK_ERR_NDT(39), WritingTestMsgError, "")
 MK_DEFINE_ERR(MK_ERR_NDT(40), QueueServerBusyError, "")
 MK_DEFINE_ERR(MK_ERR_NDT(41), FormatMsgWaitingError, "")
+MK_DEFINE_ERR(MK_ERR_NDT(42), InvalidDurationError, "")
+MK_DEFINE_ERR(MK_ERR_NDT(43), InvalidSnapsDelayError, "")
+MK_DEFINE_ERR(MK_ERR_NDT(44), InvalidNumStreamsError, "")
 
 } // namespace ndt
 } // namespace mk