+++ include/measurement_kit/ooni/templates.hpp
@@ -13,7 +13,7 @@ namespace templates {
 
 void dns_query(Var<Entry> entry, dns::QueryType, dns::QueryClass,
                std::string query_name, std::string nameserver,
-               Callback<Error, dns::Message>, Settings = {},
+               Callback<Error, Var<dns::Message>>, Settings = {},
                Var<Reactor> = Reactor::global(),
                Var<Logger> = Logger::global());