+++ src/libmeasurement_kit/libevent/ssl_context.hpp
@@ -1,8 +1,8 @@
 // Part of measurement-kit <https://measurement-kit.github.io/>.
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
-#ifndef SRC_LIBMEASUREMENT_KIT_NET_SSL_CONTEXT_HPP
-#define SRC_LIBMEASUREMENT_KIT_NET_SSL_CONTEXT_HPP
+#ifndef SRC_LIBMEASUREMENT_KIT_LIBEVENT_SSL_CONTEXT_HPP
+#define SRC_LIBMEASUREMENT_KIT_LIBEVENT_SSL_CONTEXT_HPP
 
 #include <measurement_kit/common.hpp>
 
@@ -13,26 +13,24 @@ struct ssl_st;
 struct ssl_ctx_st;
 
 namespace mk {
-namespace net {
+namespace libevent {
 
 class SslContext : public NonCopyable, public NonMovable {
   public:
-    SSL *get_client_ssl(std::string hostname);
+    ErrorOr<SSL *> get_client_ssl(std::string hostname);
 
     ~SslContext();
 
-    SslContext(std::string ca_bundle_path);
-
-    SslContext();
-
-    static Var<SslContext> global();
+    static ErrorOr<Var<SslContext>> make(std::string ca_bundle_path);
 
   private:
-    void init(std::string ca_bundle_path);
+    SslContext() {}
+    Error init(std::string ca_bundle_path);
 
+    std::string ca_bundle_path; // Used to decide whether to recreate singleton
     ssl_ctx_st *ctx = nullptr;
 };
 
-} // namespace net
+} // namespace libevent
 } // namespace mk
 #endif