+++ src/libmeasurement_kit/ndt/protocol_impl.hpp
@@ -72,7 +72,7 @@ void recv_and_ignore_kickoff_impl(Var<Context> ctx, Callback<Error> callback) {
 template <MK_MOCK_NAMESPACE(messages, read_msg),
           MK_MOCK_NAMESPACE(messages, format_msg_waiting),
           MK_MOCK_NAMESPACE(messages, write_noasync),
-          MK_MOCK(reactor_call_soon)>
+          MK_MOCK(call_soon)>
 void wait_in_queue_impl(Var<Context> ctx, Callback<Error> callback) {
     ctx->logger->debug("ndt: wait in queue ...");
     messages_read_msg(ctx, [=](Error err, uint8_t type, std::string s) {
@@ -116,12 +116,12 @@ void wait_in_queue_impl(Var<Context> ctx, Callback<Error> callback) {
                                   *wait_time);
             }
             // TODO: in theory the server can keep us in queue forever...
-            reactor_call_soon(ctx->reactor, [=]() {
+            call_soon([=]() {
                 wait_in_queue_impl<messages_read_msg,
                                    messages_format_msg_waiting,
                                    messages_write_noasync,
-                                   reactor_call_soon>(ctx, callback);
-            });
+                                   call_soon>(ctx, callback);
+            }, ctx->reactor);
             return;
         }
         ctx->logger->info("Authorized to run the test");
@@ -143,6 +143,7 @@ void recv_version_impl(Var<Context> ctx, Callback<Error> callback) {
             return;
         }
         ctx->logger->info("Got server version: %s", s.c_str());
+        (*ctx->entry)["server_version"] = s;
         // TODO: validate the server version?
         callback(NoError());
     }, ctx->reactor);
@@ -190,7 +191,7 @@ void run_tests_impl(Var<Context> ctx, Callback<Error> callback) {
         func = test_c2s_run;
     } else if (*num == TEST_META) {
         func = test_meta_run;
-    } else if (*num == TEST_S2C) {
+    } else if (*num == TEST_S2C or *num == TEST_S2C_EXT) {
         func = test_s2c_run;
     } else {
         /* nothing */