+++ src/libmeasurement_kit/net/connect.cpp
@@ -2,10 +2,17 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
+#ifdef HAVE_CONFIG_H
+#include "config.h" // For MK_CA_BUNDLE
+#endif
+
 #include "../net/connect_impl.hpp"
 #include "../net/emitter.hpp"
 #include "../net/socks5.hpp"
-#include "../net/ssl-context.hpp"
+
+#include "../libevent/connection.hpp"
+#include "../libevent/ssl_context.hpp"
+
 #include <cassert>
 #include <event2/bufferevent_ssl.h>
 #include <measurement_kit/dns.hpp>
@@ -31,7 +38,9 @@ void mk_bufferevent_on_event(bufferevent *bev, short what, void *ptr) {
 namespace mk {
 namespace net {
 
-void connect_first_of(std::vector<std::string> addresses, int port,
+using namespace mk::libevent;
+
+void connect_first_of(Var<ConnectResult> result, int port,
                       ConnectFirstOfCb cb, Settings settings,
                       Var<Reactor> reactor, Var<Logger> logger, size_t index,
                       Var<std::vector<Error>> errors) {
@@ -39,22 +48,23 @@ void connect_first_of(std::vector<std::string> addresses, int port,
     if (!errors) {
         errors.reset(new std::vector<Error>());
     }
-    if (index >= addresses.size()) {
+    if (index >= result->resolve_result.addresses.size()) {
         logger->debug("connect_first_of all addresses failed");
         cb(*errors, nullptr);
         return;
     }
     double timeout = settings.get("net/timeout", 30.0);
-    connect_base(addresses[index], port,
-                 [=](Error err, bufferevent *bev) {
+    connect_base(result->resolve_result.addresses[index], port,
+                 [=](Error err, bufferevent *bev, double connect_time) {
                      errors->push_back(err);
                      if (err) {
                          logger->debug("connect_first_of failure");
-                         connect_first_of(addresses, port, cb, settings,
+                         connect_first_of(result, port, cb, settings,
                                           reactor, logger, index + 1, errors);
                          return;
                      }
                      logger->debug("connect_first_of success");
+                     result->connect_time = connect_time;
                      cb(*errors, bev);
                  },
                  timeout, reactor, logger);
@@ -92,24 +102,24 @@ void resolve_hostname(std::string hostname, ResolveHostnameCb cb,
     logger->debug("resolve_hostname: ipv4...");
 
     dns::query("IN", "A", hostname,
-               [=](Error err, dns::Message resp) {
+               [=](Error err, Var<dns::Message> resp) {
                    logger->debug("resolve_hostname: ipv4... done");
                    result->ipv4_err = err;
-                   result->ipv4_reply = resp;
                    if (!err) {
-                       for (dns::Answer answer : resp.answers) {
+                       result->ipv4_reply = *resp;
+                       for (dns::Answer answer : resp->answers) {
                            result->addresses.push_back(answer.ipv4);
                        }
                    }
                    logger->debug("resolve_hostname: ipv6...");
                    dns::query(
                        "IN", "AAAA", hostname,
-                       [=](Error err, dns::Message resp) {
+                       [=](Error err, Var<dns::Message> resp) {
                            logger->debug("resolve_hostname: ipv6... done");
                            result->ipv6_err = err;
-                           result->ipv6_reply = resp;
                            if (!err) {
-                               for (dns::Answer answer : resp.answers) {
+                               result->ipv6_reply = *resp;
+                               for (dns::Answer answer : resp->answers) {
                                    result->addresses.push_back(answer.ipv6);
                                }
                            }
@@ -135,7 +145,7 @@ void connect_logic(std::string hostname, int port,
                          }
 
                          connect_first_of(
-                             result->resolve_result.addresses, port,
+                             result, port,
                              [=](std::vector<Error> e, bufferevent *b) {
                                  result->connect_result = e;
                                  result->connected_bev = b;
@@ -236,7 +246,7 @@ void connect(std::string address, int port,
         return;
     }
     if (settings.find("net/timeout") == settings.end()) {
-        settings["net/timeout"] = 5.0;
+        settings["net/timeout"] = 30.0;
     }
     double timeout = settings["net/timeout"].as<double>();
     connect_logic(
@@ -248,17 +258,32 @@ void connect(std::string address, int port,
                 return;
             }
             if (settings.find("net/ssl") != settings.end()) {
-                Var<SslContext> ssl_context;
+                std::string cbp
+#ifdef MK_CA_BUNDLE
+                    = MK_CA_BUNDLE
+#endif
+                ;
                 if (settings.find("net/ca_bundle_path") != settings.end()) {
-                    logger->debug("ssl: using custom ca_bundle_path");
-                    ssl_context = Var<SslContext>(
-                        new SslContext(settings.at("net/ca_bundle_path")));
-                } else {
-                    logger->debug("ssl: using default context");
-                    ssl_context = SslContext::global();
+                    cbp = settings.at("net/ca_bundle_path");
                 }
-                connect_ssl(r->connected_bev,
-                            ssl_context->get_client_ssl(address), address,
+                logger->debug("ca_bundle_path: %s", cbp.c_str());
+                ErrorOr<Var<SslContext>> ssl_context = SslContext::make(cbp);
+                if (!ssl_context) {
+                    Error err = ssl_context.as_error();
+                    err.context = r;
+                    bufferevent_free(r->connected_bev);
+                    callback(err, nullptr);
+                    return;
+                }
+                ErrorOr<SSL *> cssl = (*ssl_context)->get_client_ssl(address);
+                if (!cssl) {
+                    Error err = cssl.as_error();
+                    err.context = r;
+                    bufferevent_free(r->connected_bev);
+                    callback(err, nullptr);
+                    return;
+                }
+                connect_ssl(r->connected_bev, *cssl, address,
                             [r, callback, timeout, ssl_context, reactor,
                              logger](Error err, bufferevent *bev) {
                                 if (err) {
@@ -267,7 +292,8 @@ void connect(std::string address, int port,
                                     return;
                                 }
                                 Var<Transport> txp =
-                                    Connection::make(bev, reactor, logger);
+                                    libevent::Connection::make(
+                                        bev, reactor, logger);
                                 txp->set_timeout(timeout);
                                 assert(err == NoError());
                                 err.context = r;
@@ -277,13 +303,33 @@ void connect(std::string address, int port,
                 return;
             }
             Var<Transport> txp =
-                Connection::make(r->connected_bev, reactor, logger);
+                libevent::Connection::make(r->connected_bev, reactor, logger);
             txp->set_timeout(timeout);
             assert(err == NoError());
             err.context = r;
             callback(err, txp);
         },
         settings, reactor, logger);
+}
+
+ErrorOr<double> get_connect_time(Error err) {
+    Var<ConnectResult> cr = err.context.as<ConnectResult>();
+    if (!cr) {
+        return GenericError();
+    }
+    return cr->connect_time;
+}
+
+ErrorOr<std::vector<double>> get_connect_times(Error err) {
+    Var<ConnectManyResult> cmr = err.context.as<ConnectManyResult>();
+    if (!cmr) {
+        return GenericError();
+    }
+    std::vector<double> connect_times;
+    for (auto &cr: cmr->results) {
+        connect_times.push_back(cr->connect_time);
+    }
+    return connect_times;
 }
 
 } // namespace net