+++ src/libmeasurement_kit/ooni/ooni_test.cpp
@@ -31,7 +31,8 @@ void OoniTest::run_next_measurement(size_t thread_id, Callback<Error> cb,
         prog = *current_entry / (double)num_entries;
     }
     *current_entry += 1;
-    logger->log(MK_LOG_INFO|MK_LOG_JSON, "{\"progress\": %f}", prog);
+    logger->log(MK_LOG_INFO|MK_LOG_JSON,
+        "{\"progress\": %f, \"type\": \"progress\"}", prog);
 
     logger->debug("net_test: creating entry");
     struct tm measurement_start_time;
@@ -57,7 +58,7 @@ void OoniTest::run_next_measurement(size_t thread_id, Callback<Error> cb,
 
         report.fill_entry(entry);
         if (entry_cb) {
-            entry_cb(entry.dump());
+            entry_cb(entry.dump(4));
         }
         report.write_entry(entry, [=](Error error) {
             if (error) {
@@ -165,10 +166,10 @@ void OoniTest::open_report(Callback<Error> callback) {
     if (output_filepath == "") {
         output_filepath = generate_output_filepath();
     }
-    if (options.find("no_file_report") == options.end()) {
+    if (!options.get("no_file_report", false)) {
         report.add_reporter(FileReporter::make(output_filepath));
     }
-    if (options.find("no_collector") == options.end()) {
+    if (!options.get("no_collector", false)) {
         report.add_reporter(OoniReporter::make(*this));
     }
     report.open(callback);
@@ -184,7 +185,7 @@ std::string OoniTest::generate_output_filepath() {
         char timestamp[100];
         strftime(timestamp, sizeof(timestamp), "%FT%H%M%SZ", &test_start_time);
         filename << "report-" << test_name << "-";
-        filename << timestamp << "-" << idx << ".json";
+        filename << timestamp << "-" << idx << ".njson";
 
         std::ifstream output_file(filename.str().c_str());
         // If a file called this way already exists we increment the counter